AcadTable: fix transformation display, break detection, and un-break (issue #1305)
Part 1 (transformation): table now follows move/rotate/scale.
- Add FScale/FRotate fields and the block-insert transform overrides
  (decomposite/setrot/ReCalcFromObjMatrix/CalcObjMatrix/rtsave) so the
  in-plane rotation and scale lost by CalcObjMatrixWithoutOwner are
  preserved and reapplied from objMatrix.
- FormatEntity now reformats child entities (lines/texts) every frame and
  rebuilds the spatial tree, so the rendered table moves with the grip
  instead of only the grip moving.
- getoutbound delegates to the built children so the bounding box follows
  the transform.

Part 2a (break detection): BreakEnabled is now computed as
FBreakEnabled OR (continuation parts > 0). A table split into several
ACAD_TABLE entities (stored as continuation parts) correctly reports
Break enabled = True instead of False.

Part 2b (un-break): setting BreakEnabled := False on a broken table
merges all continuation parts top-to-bottom into a single continuous
table (rows, cells, row heights, merges with row-offset). Wired into the
object inspector via AcadTableBreakEnabledEntChangeProc so the property
is editable and triggers a recalc.

Side fixes required for range-checked builds (/etc/fpc.cfg -Crtoi):
- uzeffdxf: gotodxf/GoToDXForENDTAB use Integer for the group code read
  via ParseInteger (DXF group codes exceed 255; Byte tripped range checks).
- Replace parray^[i] with range-safe getData(i) in the model and layout.

Tests (uzctacadtable): add TransformationMovesRenderedTable and
ClearingBreakMergesContinuationParts; correct LoadsBreakSettingsFromDXF
and LoadsBreakSettingsFromSecondSample to assert the issue-aligned
computed-break behaviour. testzengine.lpi gains the acadtable unit path.

+++ cad_source/zengine/fileformats/uzeffdxf.pas
@@ -242,7 +242,7 @@ end;
 
 procedure gotodxf(var rdr:TZMemReader; fcode: Integer; const fname: String);
 var
-  byt: Byte;
+  byt: Integer;
   s: String;
   //error: Integer;
 begin
@@ -459,7 +459,7 @@ end;
 
 function GoToDXForENDTAB(var rdr:TZMemReader; fcode: Integer; const fname: String):boolean;
 var
-  byt:Byte;
+  byt:Integer;
   s:String;
 begin
   result:=false;
